# docs/openapi.yml
openapi: 3.0.3
info:
  title: VRChat Calendar API
  description: VRChatイベント募集Webアプリケーション用のREST API
  version: 1.0.0
servers:
  - url: http://localhost:3000/api
    description: ローカル開発環境
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    User:
      type: object
      properties:
        id:
          type: integer
          description: ユーザーID
          example: 1
        nickname:
          type: string
          description: ニックネーム
          example: "VRユーザー太郎"
        description:
          type: string
          description: 自己紹介
          example: "VRChat初心者です。よろしくお願いします！"
        email:
          type: string
          format: email
          description: メールアドレス
          example: "user@example.com"
        avatar_image_url:
          type: string
          description: アバター画像URL
          example: "https://example.com/avatar/user1.png"
        created_at:
          type: string
          format: date-time
          description: 作成日時
        updated_at:
          type: string
          format: date-time
          description: 更新日時
...
    Error:
      type: object
      properties:
        statusCode:
          type: integer
          example: 400
        message:
          type: string
          example: "エラーメッセージ"
        errors:
          type: array
          items:
            type: object
            properties:
              field:
                type: string
              message:
                type: string
paths:
...
  /auth/login:
    post:
      summary: ログイン
      description: ユーザーがログインします
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - nickname
                - password
              properties:
                nickname:
                  type: string
                  description: ニックネーム
                  example: "VRユーザー太郎"
                password:
                  type: string
                  description: パスワード
                  example: "password123"
      responses:
        '200':
          description: ログイン成功
          content:
            application/json:
              schema:
                type: object
                properties:
                  statusCode:
                    type: integer
                    example: 200
                  message:
                    type: string
                    example: "ログインしました"
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
                      token:
                        type: string
                        description: 認証トークン
                        example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        '401':
          description: 認証エラー
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...